fix proxy and more bugs api

+++ src/config.rs
@@ -8,7 +8,7 @@ pub struct Config {
     pub signal_server_port: i64,
     pub storage_size: u64,
     pub proxy_ip: String,
-    pub proxy_port: i64,
+    pub proxy_port: u16,
 }
 
 impl Config {
@@ -20,7 +20,7 @@ impl Config {
             signal_server_port: config["signal_server_port"].as_integer().unwrap(),
             storage_size: config["storage_size"].as_integer().unwrap() as u64,
             proxy_ip: config["proxy_ip"].as_str().unwrap().to_string(),
-            proxy_port: config["proxy_port"].as_integer().unwrap(),
+            proxy_port: config["proxy_port"].as_integer().unwrap() as u16,
         }
     }
 }
